requirements.txt
flask_socketio
flask
eventlet
opencv-python
opencv-contrib-python
numpy
pillow